Setup Travis CI builds
This commit enables Travis CI continuous integration builds at
<https://travis-ci.org/>.

A build matrix is defined to create debug and production builds
with libgcrypt and openssl cryptographic libraries using gcc and
clang compilers.

colormake is used to colorize the output of the Travis CI job log.

The "-Werror" CFLAGS option is specified for all builds in order
to change warnings into errors.  Note that neither "-pedantic" nor
"-pedantic-errors" are specified in CFLAGS.  libtool will fail if
either is used in the CFLAGS passed to configure.  It should be
possible to override libtool's CFLAGS with LTCFLAGS, but this does
not work; LTCFLAGS always takes on the value of CFLAGS.  Grrr.

Reference:
- https://docs.travis-ci.com/user/getting-started/

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>