Specify C99 language standard in debug build
The gcc version in Ubuntu 12.04 (used for the standard Travis CI
virtualization environment) defaults to the C90 language standard.
This causes "-pedantic" builds to fail with the following errors:

> unmunge.c: In function 'display_meta':
> unmunge.c:627:17: error: ISO C90 does not support the '%z'
>       gnu_strftime format [-Werror=format]
> unmunge.c:651:17: error: ISO C90 does not support the '%z'
>       gnu_strftime format [-Werror=format]

This commit updates the DEBUGCFLAGS specified via "--enable-debug"
to use the C99 language standard.  C99 appears to support the "%z"
strftime format arg.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>